Prevent node wrapper mixups, fix nullpointer.
Without clearing the node registry, there could be nodes that match a
relation of an item in the current search result list. This would lead to
that node being treated as a proper search result tree entry, but since
its object doesn't occur in the current search result list, it can't be
handled by the sort comparator. This can be fixed by clearing the node
registry and making sure there is no nullpointer when looking up a
non-existent search result list position.