Allow filtering of lemma search dialog search results.
It is desirable to be able to apply a person names filter mechanism
similar to the one active in the automatic lemmatizer search even if the
search results were queried via the general search dialog.

A nice solution would be to save the most previous search query so that
it can be replayed whenever the 'include persons' flag is being toggled
by the user, and just apply the exclusion clause to the query when
appropriate. Unfortunately, we would need a deep copy of it for this,
and the `BTSQueryRequest` is very much uncloneable. (It is in dire need
of a makeover anyway there's too much state inconsistencies and it's
become kinda bloated)

So instead there's just a viewer filter now which supresses person names
if the flag is set.